refactor : 컨트롤러, 서비스 파일 통합하기
