+++ src/main/java/com/opus/opus/modules/team/application/TeamCommandService.java
@@ -6,14 +6,26 @@ import static com.opus.opus.modules.file.domain.FileImageType.THUMBNAIL;
 import static com.opus.opus.modules.file.domain.ReferenceDomainType.TEAM;
 import static com.opus.opus.modules.file.exception.FileExceptionType.EXCEED_PREVIEW_LIMIT;
 import static com.opus.opus.modules.file.exception.FileExceptionType.NOT_WEBP_CONVERTED;
+import static com.opus.opus.modules.team.exception.TeamLikeExceptionType.ALREADY_LIKED;
+import static com.opus.opus.modules.team.exception.TeamLikeExceptionType.ALREADY_UNLIKED;
+import static com.opus.opus.modules.team.exception.TeamLikeExceptionType.NOT_LIKED_YET;
 
 import com.opus.opus.global.util.FileStorageUtil;
+import com.opus.opus.modules.contest.application.convenience.ContestConvenience;
+import com.opus.opus.modules.contest.domain.Contest;
 import com.opus.opus.modules.file.domain.File;
 import com.opus.opus.modules.file.domain.FileImageType;
 import com.opus.opus.modules.file.domain.dao.FileRepository;
 import com.opus.opus.modules.file.exception.FileException;
 import com.opus.opus.modules.team.application.convenience.TeamConvenience;
+import com.opus.opus.modules.team.application.dto.response.TeamLikeToggleResponse;
+import com.opus.opus.modules.team.domain.Team;
+import com.opus.opus.modules.team.domain.TeamLike;
+import com.opus.opus.modules.team.domain.dao.TeamLikeRepository;
+import com.opus.opus.modules.team.exception.TeamLikeException;
 import java.util.List;
+import java.util.Objects;
+import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -28,6 +40,9 @@ public class TeamCommandService {
 
     private final FileRepository fileRepository;
     private final TeamConvenience teamConvenience;
+    private final ContestConvenience contestConvenience;
+
+    private final TeamLikeRepository teamLikeRepository;
 
 
     public void savePreviewImages(final Long teamId, final List<MultipartFile> images) {
@@ -63,6 +78,45 @@ public class TeamCommandService {
     public void deletePosterImage(final Long teamId) {
         teamConvenience.validateExistTeam(teamId);
         deleteIfExists(teamId, POSTER);
+    }
+
+    public TeamLikeToggleResponse toggleLike(final Long memberId, final Long teamId, final Boolean isLiked) {
+        final Team team = teamConvenience.getValidateExistTeam(teamId);
+        final Contest contest = contestConvenience.getValidateExistContest(team.getContestId());
+
+        contestConvenience.validateNotInVotingPeriod(contest);
+
+        final Optional<TeamLike> teamLikeOptional = teamLikeRepository.findByMemberIdAndTeam(memberId, team);
+        return teamLikeOptional
+                .map(teamLike -> handleExistingLike(teamLike, isLiked))
+                .orElseGet(() -> handleFirstTimeLike(memberId, team, isLiked));
+    }
+
+    private TeamLikeToggleResponse handleFirstTimeLike(final Long memberId, final Team team, final Boolean isLiked) {
+        if (!isLiked) {
+            throw new TeamLikeException(NOT_LIKED_YET);
+        }
+
+        saveTeamLike(memberId, team);
+        return TeamLikeToggleResponse.of(team.getId(), true, "좋아요가 등록되었습니다.");
+    }
+
+    private TeamLikeToggleResponse handleExistingLike(final TeamLike teamLike, final Boolean isLiked) {
+        if (Objects.equals(teamLike.getIsLiked(), isLiked)) {
+            throw new TeamLikeException(isLiked ? ALREADY_LIKED : ALREADY_UNLIKED);
+        }
+
+        teamLike.updateIsLiked(isLiked);
+
+        return TeamLikeToggleResponse.of(teamLike.getTeam().getId(), isLiked, isLiked ? "좋아요가 등록되었습니다." : "좋아요가 취소되었습니다.");
+    }
+
+    private void saveTeamLike(final Long memberId, final Team team) {
+        teamLikeRepository.save(TeamLike.builder()
+                .memberId(memberId)
+                .team(team)
+                .isLiked(true)
+                .build());
     }
 
     private void deleteIfExists(final Long teamId, final FileImageType imageType) {
